+++ js/get-data.js
@@ -1,21 +1,23 @@
-import {closeUploadModal} from './form-control.js';
+import {closeForm} from './modal-control.js';
 import {openBigPictureModal} from './big-picture.js';
 import {renderSmallPictures} from './small-pictures.js';
 import {formSubmit} from './send-data.js';
 
 const filtersContainer = document.querySelector('.img-filters');
 
-const getData = () => {
+const getData = (onSuccess) => {
   fetch('https://22.javascript.pages.academy/kekstagram/data')
     .then((response) => response.json())
     .then((data) => {
-      renderSmallPictures(data);
-      openBigPictureModal(data);
-      filtersContainer.classList.remove('img-filters--inactive');
+      onSuccess(data)
     });
 }
 
-getData();
-formSubmit(closeUploadModal);
+getData((data) => {
+  renderSmallPictures(data);
+  openBigPictureModal(data);
+  filtersContainer.classList.remove('img-filters--inactive');
+});
 
+formSubmit(closeForm);
 export {getData};
